:ok_hand: [#1182] PR feedback
